requirements.txt
opencv-python>=4.5.0
torch>=1.12.0
tensorflow>=2.10.0
numpy>=1.21.0
pyyaml>=6.0.0